Rename `mining` mod to `bitcoin_mining`
