+++ testnet/stacks-node/src/tests/signer/commands/bitcoin_mining.rs
