+++ testnet/stacks-node/src/tests/signer/commands/mod.rs
@@ -1,21 +1,21 @@
 mod context;
 
+mod bitcoin_mining;
 mod block_commit;
 mod block_wait;
 mod boot;
 mod commit_ops;
-mod mining;
 mod shutdown;
 mod stacks_mining;
 mod transfer;
 
 pub use context::SignerTestContext;
 
+pub use bitcoin_mining::{MineBitcoinBlock, MineBitcoinBlockTenureChangePrimaryMiner};
 pub use block_commit::{SubmitBlockCommitPrimaryMiner, SubmitBlockCommitSecondaryMiner};
 pub use block_wait::{WaitForBlockFromMiner1, WaitForBlockFromMiner2};
 pub use boot::BootToEpoch3;
 pub use commit_ops::{SkipCommitOpPrimaryMiner, SkipCommitOpSecondaryMiner};
-pub use mining::{MineBitcoinBlock, MineBitcoinBlockTenureChangePrimaryMiner};
 pub use shutdown::ShutdownMiners;
 pub use stacks_mining::{PauseStacksMining, ResumeStacksMining};
 pub use transfer::SendTransferTx;
